+++ Rx.playground/Pages/Filtering_and_Conditional_Operators.xcplaygroundpage/Contents.swift
@@ -74,6 +74,16 @@ example("single with conditions") {
     let disposeBag = DisposeBag()
     
     Observable.of("🐱", "🐰", "🐶", "🐸", "🐷", "🐵")
+        .single { $0 == "🐸" }
+        .subscribe { print($0) }
+        .addDisposableTo(disposeBag)
+    
+    Observable.of("🐱", "🐰", "🐶", "🐱", "🐰", "🐶")
+        .single { $0 == "🐰" }
+        .subscribe { print($0) }
+        .addDisposableTo(disposeBag)
+    
+    Observable.of("🐱", "🐰", "🐶", "🐸", "🐷", "🐵")
         .single { $0 == "🔵" }
         .subscribe { print($0) }
         .addDisposableTo(disposeBag)