+++ RxCocoa/Common/DelegateProxyType.swift
@@ -178,13 +178,13 @@ extension DelegateProxyType {
         let maybeProxy = Self.assignedProxyFor(object) as? Self
 
         let proxy: Self
-        if maybeProxy == nil {
+        if let existingProxy = maybeProxy {
+            proxy = existingProxy
+        }
+        else {
             proxy = Self.createProxyForObject(object) as! Self
             Self.assignProxy(proxy, toObject: object)
             assert(Self.assignedProxyFor(object) === proxy)
-        }
-        else {
-            proxy = maybeProxy!
         }
 
         let currentDelegate: AnyObject? = Self.currentDelegateFor(object)
@@ -247,7 +247,11 @@ extension ObservableType {
         let disposable = P.installForwardDelegate(dataSource, retainDelegate: retainDataSource, onProxyForObject: object)
 
         let subscription = self.asObservable()
-            // source can never end, otherwise it would release the subscriber
+            .catchError { error in
+                bindingErrorToInterface(error)
+                return Observable.empty()
+            }
+            // source can never end, otherwise it would release the subscriber, and deallocate the data source
             .concat(Observable.never())
             .subscribe { [weak object] (event: Event<E>) in
                 MainScheduler.ensureExecutingOnScheduler()
@@ -269,6 +273,6 @@ extension ObservableType {
                 }
             }
             
-        return StableCompositeDisposable.create(subscription, disposable)
+        return CompositeDisposable(subscription, disposable)
     }
 }