+++ RxCocoa/iOS/UITableView+Rx.swift
@@ -24,6 +24,23 @@ extension UITableView {
     - parameter source: Observable sequence of items.
     - parameter cellFactory: Transform between sequence elements and view cells.
     - returns: Disposable object that can be used to unbind.
+     
+     Example:
+    
+         let items = Observable.just([
+             "First Item",
+             "Second Item",
+             "Third Item"
+         ])
+
+         items
+         .bindTo(tableView.rx_itemsWithCellFactory) { (tableView, row, element) in
+             let cell = tableView.dequeueReusableCellWithIdentifier("Cell")!
+             cell.textLabel?.text = "\(element) @ row \(row)"
+             return cell
+         }
+         .addDisposableTo(disposeBag)
+
     */
     public func rx_itemsWithCellFactory<S: SequenceType, O: ObservableType where O.E == S>
         (source: O)
@@ -44,6 +61,20 @@ extension UITableView {
     - parameter configureCell: Transform between sequence elements and view cells.
     - parameter cellType: Type of table view cell.
     - returns: Disposable object that can be used to unbind.
+     
+     Example:
+
+         let items = Observable.just([
+             "First Item",
+             "Second Item",
+             "Third Item"
+         ])
+
+         items
+             .bindTo(tableView.rx_itemsWithCellIdentifier("Cell", cellType: UITableViewCell.self)) { (row, element, cell) in
+                cell.textLabel?.text = "\(element) @ row \(row)"
+             }
+             .addDisposableTo(disposeBag)
     */
     public func rx_itemsWithCellIdentifier<S: SequenceType, Cell: UITableViewCell, O : ObservableType where O.E == S>
         (cellIdentifier: String, cellType: Cell.Type = Cell.self)
@@ -65,10 +96,46 @@ extension UITableView {
     
     /**
     Binds sequences of elements to table view rows using a custom reactive data used to perform the transformation.
+    This method will retain the data source for as long as the subscription isn't disposed (result `Disposable` 
+    being disposed).
+    In case `source` observable sequence terminates sucessfully, the data source will present latest element
+    until the subscription isn't disposed.
     
     - parameter dataSource: Data source used to transform elements to view cells.
     - parameter source: Observable sequence of items.
     - returns: Disposable object that can be used to unbind.
+     
+     Example 
+
+        let dataSource = RxTableViewSectionedReloadDataSource<SectionModel<String, Double>>()
+
+        let items = Observable.just([
+            SectionModel(model: "First section", items: [
+                1.0,
+                2.0,
+                3.0
+                ]),
+            SectionModel(model: "Second section", items: [
+                1.0,
+                2.0,
+                3.0
+                ]),
+            SectionModel(model: "Third section", items: [
+                1.0,
+                2.0,
+                3.0
+                ])
+            ])
+
+        dataSource.configureCell = { (dataSource, tv, indexPath, element) in
+        let cell = tv.dequeueReusableCellWithIdentifier("Cell")!
+            cell.textLabel?.text = "\(element) @ row \(indexPath.row)"
+            return cell
+        }
+
+        items
+            .bindTo(tableView.rx_itemsWithDataSource(dataSource))
+            .addDisposableTo(disposeBag)
     */
     public func rx_itemsWithDataSource<
             DataSource: protocol<RxTableViewDataSourceType, UITableViewDataSource>,
@@ -78,7 +145,15 @@ extension UITableView {
         -> (source: O)
         -> Disposable  {
         return { source in
-            return source.subscribeProxyDataSourceForObject(self, dataSource: dataSource, retainDataSource: false) { [weak self] (_: RxTableViewDataSourceProxy, event) -> Void in
+            // This is called for sideeffects only, and to make sure delegate proxy is in place when
+            // data source is being bound.
+            // This is needed because theoretically the data source subscription itself might
+            // call `self.rx_delegate`. If that happens, it might cause weird side effects since
+            // setting data source will set delegate, and UITableView might get into a weird state.
+            // Therefore it's better to set delegate proxy first, just to be sure.
+            _ = self.rx_delegate
+            // Strong reference is needed because data source is in use until result subscription is disposed
+            return source.subscribeProxyDataSourceForObject(self, dataSource: dataSource, retainDataSource: true) { [weak self] (_: RxTableViewDataSourceProxy, event) -> Void in
                 guard let tableView = self else {
                     return
                 }
@@ -119,9 +194,10 @@ extension UITableView {
    
     /**
     Installs data source as forwarding delegate on `rx_dataSource`.
+    Data source won't be retained.
     
     It enables using normal delegate mechanism with reactive delegate mechanism.
-    
+     
     - parameter dataSource: Data source object.
     - returns: Disposable object that can be used to unbind the data source.
     */
@@ -214,7 +290,7 @@ extension UITableView {
     Reactive wrapper for `delegate` message `tableView:willDisplayCell:forRowAtIndexPath:`.
     */
     public var rx_willDisplayCell: ControlEvent<WillDisplayCellEvent> {
-        let source: Observable<DidEndDisplayingCellEvent> = rx_delegate.observe(#selector(UITableViewDelegate.tableView(_:willDisplayCell:forRowAtIndexPath:)))
+        let source: Observable<WillDisplayCellEvent> = rx_delegate.observe(#selector(UITableViewDelegate.tableView(_:willDisplayCell:forRowAtIndexPath:)))
             .map { a in
                 return (try castOrThrow(UITableViewCell.self, a[1]), try castOrThrow(NSIndexPath.self, a[2]))
             }