+++ RxExample/RxDataSources/DataSources+Rx/RxCollectionViewSectionedAnimatedDataSource.swift
@@ -29,6 +29,9 @@ public class RxCollectionViewSectionedAnimatedDataSource<S: AnimatableSectionMod
     
     public func collectionView(collectionView: UICollectionView, observedEvent: Event<Element>) {
         UIBindingObserver(UIElement: self) { dataSource, newSections in
+            #if DEBUG
+                self._dataSourceBound = true
+            #endif
             if !self.dataSet {
                 self.dataSet = true
                 dataSource.setSections(newSections)
@@ -47,7 +50,10 @@ public class RxCollectionViewSectionedAnimatedDataSource<S: AnimatableSectionMod
                         }
                     }
                     catch let e {
+                        #if DEBUG
+                        print("Error while binding data animated: \(e)\nFallback to normal `reloadData` behavior.")
                         rxDebugFatalError(e)
+                        #endif
                         self.setSections(newSections)
                         collectionView.reloadData()
                     }