+++ RxExample/RxDataSources/DataSources+Rx/RxTableViewSectionedAnimatedDataSource.swift
@@ -28,6 +28,9 @@ public class RxTableViewSectionedAnimatedDataSource<S: AnimatableSectionModelTyp
 
     public func tableView(tableView: UITableView, observedEvent: Event<Element>) {
         UIBindingObserver(UIElement: self) { dataSource, newSections in
+            #if DEBUG
+                self._dataSourceBound = true
+            #endif
             if !self.dataSet {
                 self.dataSet = true
                 dataSource.setSections(newSections)