+++ RxExample/RxDataSources/DataSources/Differentiator.swift
@@ -13,6 +13,7 @@ public enum DifferentiatorError
     , CustomDebugStringConvertible {
     case DuplicateItem(item: Any)
     case DuplicateSection(section: Any)
+    case InvalidInitializerImplementation(section: Any, expectedItems: Any, expectedIdentifier: Any)
 }
 
 extension DifferentiatorError {
@@ -22,6 +23,10 @@ extension DifferentiatorError {
             return "Duplicate item \(item)"
         case let .DuplicateSection(section):
             return "Duplicate section \(section)"
+        case let InvalidInitializerImplementation(section, expectedItems, expectedIdentifier):
+            return "Wrong initializer implementation for: \(section)\n" +
+                "Expected it should return items: \(expectedItems)\n" +
+                "Expected it should have id: \(expectedIdentifier)"
         }
     }
 }
@@ -104,7 +109,9 @@ func indexSections<S: AnimatableSectionModelType>(sections: [S]) throws -> [S.Id
     for (i, section) in sections.enumerate() {
         guard indexedSections[section.identity] == nil else {
             #if DEBUG
-            precondition(indexedSections[section.identity] == nil, "Section \(section) has already been indexed at \(indexedSections[section.identity]!)")
+                if indexedSections[section.identity] != nil {
+                    print("Section \(section) has already been indexed at \(indexedSections[section.identity]!)")
+                }
             #endif
             throw DifferentiatorError.DuplicateItem(item: section)
         }
@@ -127,7 +134,9 @@ func indexSectionItems<S: AnimatableSectionModelType>(sections: [S]) throws -> [
         for (j, item) in sections[i].items.enumerate() {
             guard indexedItems[item.identity] == nil else {
                 #if DEBUG
-                precondition(indexedItems[item.identity] == nil, "Item \(item) has already been indexed at \(indexedItems[item.identity]!)" )
+                    if indexedItems[item.identity] != nil {
+                        print("Item \(item) has already been indexed at \(indexedItems[item.identity]!)" )
+                    }
                 #endif
                 throw DifferentiatorError.DuplicateItem(item: item)
             }
@@ -278,6 +287,16 @@ public func differencesForSectionedView<S: AnimatableSectionModelType>(
     result.appendContentsOf(try sectionCommands.generateNewAndMovedItems())
 
     return result
+}
+
+private extension AnimatableSectionModelType {
+    init(safeOriginal: Self, safeItems: [Item]) throws {
+        self.init(original: safeOriginal, items: safeItems)
+
+        if self.items != safeItems || self.identity != safeOriginal.identity {
+            throw DifferentiatorError.InvalidInitializerImplementation(section: self, expectedItems: safeItems, expectedIdentifier: safeOriginal.identity)
+        }
+    }
 }
 
 struct CommandGenerator<S: AnimatableSectionModelType> {
@@ -541,7 +560,7 @@ struct CommandGenerator<S: AnimatableSectionModelType> {
                 }
             }
 
-            afterDeleteState.append(S(original: initialSection, items: afterDeleteItems))
+            afterDeleteState.append(try S(safeOriginal: initialSection, safeItems: afterDeleteItems))
         }
         // }
 
@@ -616,7 +635,9 @@ struct CommandGenerator<S: AnimatableSectionModelType> {
                     items.append(self.finalSections[finalIndex.sectionIndex].items[finalIndex.itemIndex])
                 }
                 
-                return S(original: s, items: items)
+                let modifiedSection = try S(safeOriginal: s, safeItems: items)
+
+                return modifiedSection
             }
             else {
                 try rxPrecondition(false, "This is weird, this shouldn't happen")