+++ RxSwift/Schedulers/MainScheduler.swift
@@ -43,9 +43,9 @@ public final class MainScheduler : SerialDispatchQueueScheduler {
     /**
     In case this method is called on a background thread it will throw an exception.
     */
-    public class func ensureExecutingOnScheduler() {
+    public class func ensureExecutingOnScheduler(errorMessage: String? = nil) {
         if !NSThread.currentThread().isMainThread {
-            rxFatalError("Executing on backgound thread. Please use `MainScheduler.instance.schedule` to schedule work on main thread.")
+            rxFatalError(errorMessage ?? "Executing on backgound thread. Please use `MainScheduler.instance.schedule` to schedule work on main thread.")
         }
     }