+++ RxTests/Recorded.swift
@@ -41,7 +41,7 @@ extension Recorded {
     }
 }
 
-func == <T: Equatable>(lhs: Recorded<T>, rhs: Recorded<T>) -> Bool {
+public func == <T: Equatable>(lhs: Recorded<T>, rhs: Recorded<T>) -> Bool {
     return lhs.time == rhs.time && lhs.value == rhs.value
 }